requirements.txt
Collectfast==0.1.13
Django==1.7.1
South==0.8.4
boto==2.25.0
django-autocomplete-light==1.4.11
django-pipeline==1.3.23
django-storages==1.1.8
djangorestframework==2.3.10
futures==2.1.6
mysql-python==1.2.4b4
python-dateutil==2.2
pytz==2013.9
six==1.5.2
